backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pyrogram==2.0.106
tgcrypto==1.2.5
sqlalchemy==1.4.50
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
aiofiles==23.2.1
python-dotenv==1.0.1 